# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Planton — Plataforma de Plantões Médicos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/profile:
    post:
      operationId: completeProfile
      tags: [auth]
      summary: Completar perfil profissional após autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "201":
          description: Perfil profissional criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Dados profissionais inválidos ou já cadastrados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Identidade Google não autenticada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Perfil profissional já concluído
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /login:
    get:
      operationId: beginBrowserLogin
      tags: [auth]
      summary: Iniciar login OIDC
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirecionamento para o provedor de identidade
...
  /logout:
    get:
      operationId: logoutBrowserSession
      tags: [auth]
      summary: Encerrar sessão
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirecionamento após logout
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Usuário logado atual
      responses:
        "200":
          description: Dados do usuário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Doctor subscription ───────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── User ───────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, type, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        type:
          type: string
          enum: [doctor, hospital]
        specialty:
          type: ["string", "null"]
        crmNumber:
          type: ["string", "null"]
        crmState:
          type: ["string", "null"]
        hospitalName:
          type: ["string", "null"]
        cnpj:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        createdAt:
          type: string
        subscription:
          oneOf:
            - $ref: "#/components/schemas/SubscriptionSummary"
            - type: "null"
    SubscriptionSummary:
      type: object
      required: [id, value, cycle, status, nextDueDate, invoiceUrl, lastPaymentStatus]
      properties:
        id:
          type: number
        value:
          type: number
        cycle:
          type: string
          enum: [MONTHLY]
        status:
          type: string
          enum: [PENDENTE, ATIVA, INADIMPLENTE, CANCELADA]
        nextDueDate:
          type: ["string", "null"]
        invoiceUrl:
          type: ["string", "null"]
        lastPaymentStatus:
          type: ["string", "null"]
        lastPaymentAt:
          type: ["string", "null"]
...
    ProfileInput:
      type: object
      description: O formato completo depende do tipo de perfil.
      oneOf:
        - $ref: "#/components/schemas/DoctorProfileInput"
        - $ref: "#/components/schemas/HospitalProfileInput"
    DoctorProfileInput:
      type: object
      required: [name, type, specialty, phone, city, state]
      properties:
        name:
          type: string
          minLength: 2
        type:
          type: string
          enum: [doctor]
        specialty:
          type: string
          minLength: 1
        crmNumber:
          type: string
          minLength: 1
        crmState:
          type: string
          minLength: 2
          maxLength: 2
        phone:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 2
          maxLength: 2
    HospitalProfileInput:
      type: object
      required: [name, type, hospitalName, phone, address, city, state]
      properties:
        name:
          type: string
          minLength: 2
        type:
          type: string
          enum: [hospital]
        hospitalName:
          type: string
          minLength: 2
        cnpj:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 2
          maxLength: 2

    # ── Shift ──────────────────────────────────────────────────────────────
...