# lib/api-zod/src/generated/types/doctorProfileInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Planton — Plataforma de Plantões Médicos
 * OpenAPI spec version: 1.0.0
 */
import type { DoctorProfileInputType } from './doctorProfileInputType';

export interface DoctorProfileInput {
  /** @minLength 2 */
  name: string;
  type: DoctorProfileInputType;
  /** @minLength 1 */
  specialty: string;
  /** @minLength 1 */
  crmNumber?: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  crmState?: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  city: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  state: string;
}