# lib/api-zod/src/generated/types/hospitalProfileInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Planton — Plataforma de Plantões Médicos
 * OpenAPI spec version: 1.0.0
 */
import type { HospitalProfileInputType } from './hospitalProfileInputType';

export interface HospitalProfileInput {
  /** @minLength 2 */
  name: string;
  type: HospitalProfileInputType;
  /** @minLength 2 */
  hospitalName: string;
  /** @minLength 1 */
  cnpj?: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  city: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  state: string;
}