# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Planton — Plataforma de Plantões Médicos
 * OpenAPI spec version: 1.0.0
 */
import type { SubscriptionSummary } from './subscriptionSummary';
import type { UserType } from './userType';

export interface User {
  id: number;
  name: string;
  email: string;
  type: UserType;
  /** @nullable */
  specialty?: string | null;
  /** @nullable */
  crmNumber?: string | null;
  /** @nullable */
  crmState?: string | null;
  /** @nullable */
  hospitalName?: string | null;
  /** @nullable */
  cnpj?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  createdAt: string;
  subscription?: SubscriptionSummary | null;
}